+++ CHANGELOG.rst
@@ -17,6 +17,10 @@ Added
 
 * Added ``add_entropy_neff`` option to ``alignment_to_count_df``
 
+* Added the ``colorschemes`` module.
+
+* Added ``pymol_struct`` module to color ``pymol`` structures.
+
 Fixed
 +++++
 * Made ``--drop_pdb`` and ``--ignore_gaps`` options work correctly (before they did nothing due to command-line parsing error)